refactor to use clap build, remove loop logic, refactor u64 to usize

+++ src/timer.rs
@@ -1,9 +1,9 @@
 use std::{thread, time};
 
-pub fn timer(time_seconds: u64, numerals: Vec<Vec<String>>){
+pub fn timer(time_seconds: usize, numerals: Vec<Vec<String>>){
     // takes in time and detracts it after pausing for one second
 
-    let mut time: u64 = time_seconds;
+    let mut time: usize = time_seconds;
     let ten_millis: time::Duration = time::Duration::from_millis(1000);
     loop {
 
@@ -27,7 +27,7 @@ pub fn timer(time_seconds: u64, numerals: Vec<Vec<String>>){
 
 }
 
-pub fn pretty_display(min:u64, sec:u64, numerals: &Vec<Vec<String>>) {
+pub fn pretty_display(min:usize, sec:usize, numerals: &Vec<Vec<String>>) {
     let min = min as usize;
     let sec = sec as usize;
